feat: Implement multi-user support and enhance navigation
This commit introduces significant updates to enable a multi-user environment, primarily by replacing hardcoded user IDs with dynamic data. It also improves user experience by making user profiles and reviews accessible from various screens and refactoring UI components.

Key changes:
-   **Multi-User & Dynamic Data:**
    -   Updated the data seeder to assign catches to multiple fishers randomly and ensure offers are created with the correct associated fisher data.
    -   Modified notification screens for both Buyers and Fishers to filter content based on the logged-in user's ID, which is now passed via the route.
    -   The `HomeScreen` now correctly filters catches and builds the notification link based on the current user's ID from `UserBloc`.

-   **Navigation to Reviews:**
    -   User profile sections in `ProductDetails` (Seller) and `OrderDetails` (Buyer/Fisher) are now tappable, navigating to the respective user's review screen.
    -   Routing has been updated to handle dynamic navigation to review pages (e.g., `/buyer/reviews/:id`).

-   **UI & Component Refinement:**
    -   **`PageTitle` Widget:** Replaced hardcoded `Text` widgets for screen titles with a new, reusable `PageTitle` widget in `OrderDetails` and `NotificationsScreen`.
    -   **Offer Actions:** Added a message in the `OfferActions` widget to inform users when they are waiting for a response from the other party.
    -   **Pending Offers:** The "Make Offer" button on the `ProductDetails` screen is now disabled and displays "Offer Pending" if the current user already has an active offer on that item.
    -   **Local `animated_rating_stars`:** The `animated_rating_stars` package was removed and replaced with an identical local widget to resolve a dependency issue. Star sizes in the `RatingModalContent` were also increased for better usability.

-   **Dependency Management:**
    -   Removed the `animated_rating_stars` package from `pubspec.yaml` and `pubspec.lock`.

+++ lib/features/user/presentation/widgets/rating_card.dart
@@ -1,7 +1,7 @@
-import 'package:animated_rating_stars/animated_rating_stars.dart';
 import 'package:flutter/material.dart';
 import 'package:siren_marketplace/core/constants/app_colors.dart';
 import 'package:siren_marketplace/core/utils/custom_icons.dart';
+import 'package:siren_marketplace/core/widgets/animated_rating_stars.dart';
 
 class RatingCard extends StatelessWidget {
   final double averageRating;
@@ -66,10 +66,9 @@ class RatingCard extends StatelessWidget {
                 ),
                 const SizedBox(height: 24),
                 AnimatedRatingStars(
-                  initialRating: averageRating,
+                  initialRating: averageRating.toDouble(),
                   minRating: 1,
                   maxRating: 5.0,
-
                   filledColor: AppColors.shellOrange,
                   emptyColor: AppColors.gray100,
                   onChanged: (v) => null,
